Add Support for Java Programs

+++ src/Program/JavaProgram.php
@@ -1,22 +1,29 @@
 <?php
 namespace Manthan\Webcompile\Program;
 
+use SplFileObject;
+
 class JavaProgram extends Program {
     protected $compiler   = 'javac';
 
-    public function __construct($file, $args) {
-        Parent::__construct($file, $args);
+    public function create() {
+        $this->file = new SplFileObject($this->name, "w");
+        $this->file->fwrite($this->content);
     }
 
     protected function getRunCommand() {
-        $fileName = $this->extractFileName($this->file);
-        return 'java '.$fileName;
+        $fileName = $this->extractFileName($this->name);
+        return "java {$fileName} 2>&1";
     }
 
     public function destroy() {
-        unlink($this->file);
-        $classFile = $this->extractFileName($this->file).'.class';
-        unlink($classFile);
+        if(file_exists($this->name)) {
+         unlink($this->name);
+        }
+        $classFile = $this->extractFileName($this->name).'.class';
+        if(file_exists($classFile)) {
+         unlink($classFile);
+       }
     }
 
     protected function extractFileName($full) {
